+++ src/components/Button/index.ts
@@ -1,1 +1,2 @@
 export { default as EmptyCardButton } from './EmptyCardButton';
+export { default as TextButton } from './TextButton';
