+++ src/components/index.ts
@@ -1,2 +1,5 @@
-export { TextInput } from './TextInput';
-export { Label } from './Label';
+export * from './TextInput';
+export * from './Label';
+export * from './Button';
+export * from './Card';
+export * from './Header';
