+++ src/pages/CardListPage/CardListPage.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Header } from '@/components/Header';
-import { EmptyCardButton } from '@/components/Button';
-import { useNavigate } from 'react-router-dom';
-import CardList from './components/CardList/CardList';
+import { CardList } from './components';
 
 const CardListPage = () => {
   return (
