+++ src/pages/CardListPage/components/index.ts
@@ -1,0 +1,1 @@
+export * from './CardList';
